requirements.txt
aniso8601==4.0.1
arrow==0.12.1
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
flask-restplus==0.12.1
fuzzywuzzy==0.17.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
pybreezechms==1.0.0
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
-e git+https://github.com/jaycle/pyBreezeChMS.git@fix_error_handling#egg=pybreezechms
